Add another Sokoban level adapted from Slash'em:  Into the Middle.
I replaced the iron bars with walls, because having to repeatedly walk
past iron bars a bunch of times before you can get behind there and
finally deal with whatever ranged-attack monger generated behind the
bars is evil, especially this early in the game.

But other than that this is straight out of Slash'em, and not a
very difficult puzzle, considering.