Move wizards' guaranteed magic marker to the quest.
The magic marker for wizards started out being a startscum item in
starting inventory in vanilla; I originally changed that to a
guaranteed item.  When the Pw regen formula went in, I rebalanced
wizard starting inventory and, as part of that rebalance, moved the
magic marker to the magic chest.  This moves it further into the game,
to the quest locate level.

Moving it to the quest nemesis' inventory, as wheals suggested, and
making him spend its charges writing things to hurt you with, is left
unimplemented for the time being.