Monster difficulty (monstr) adjustments: tweak mstrength formula, and abolish it.
First I tweaked the mstrength formula in makedefs.c, to take into
account some things it wasn't taking into account, including wand
skill, 0d0 attacks (which are more dangerous in many cases than their
numbers would suggest), and domesticity.

Then I generated a new monster spoiler, using this formula, and
compared it to the former one, side-by-side.

Then, in a discussion on IRC, ais523 suggested removing the automated
monstr calculation altogether, since the formula is fundamentally
rather arbitrary and doesn't take into account that some combinations
of dangerous features are more dangerous than others.

So I removed the mstrength() function from makedefs.c, removed the
monstr[] array from the game, and added the difficulty field to the
permonst struct.  For the moment, most references to it are wrapped in
a MONSTR() macro; not sure if I want to keep that.  Ok for now.