Make setequip actually honor em_silent.
The setequip function has an argument, msgtype, which can be set to
em_silent to indicate that messages should not be generated.  It was
honored for one or two of the messages that setequip can generate but
not for most of them, probably because most of them never occurred
under the em_silent circumstances (notably, when automatically
equipping starting armor at game start) in vanilla.

This fixes the "You feel yourself slowing down.  You feel yourself
speeding up." bug in Fourk and heads off a number of potential others
in case a player ever starts with any of the other items that cause
messages when equipped.

I didn't suppress the amulet of change messages, however, because
it would probably be weirder to not get the message in that case.