Kicking secrets: be a little more player-friendly.
I don't see the point in wounding the player's legs in this situation,
causing them to just have to wait for a few dozen turns before trying
the kick again.  Really, that's just tedious and in no way makes the
game more genuinely challenging.

Thanks to Khor for pointing this out.