No more mundane peacefuls in Sokoban.
In Sokoban, monsters that would normally be generated peaceful will
instead now generate as hostile, unless they are M2_NASTY or always
peaceful.

I'm not sure if anything always peaceful generates in Sokoban, but I
coded the exception for them just in case.

In most games, I would imagine, nothing M2_NASTY will generate there
either, in the early game when the player is still trying to solve the
boulder puzzle; but if a player uses Sokoban for a stash location,
something nasty might generate there on later visits.  If so, it will
have its usual chance of being peaceful.

The reasoning behind this change is that mundane monsters are, in
Sokoban, uniquely worse for the player than if they were hostile.
Sokoban levels are small (especially for the amount of time spent
there) and easily clogged up and particularly sensitive to such
clogging, both because they're undiggable and tend to have a lot of
narrow spots and also because the player is often attempting to push
boulders around in very specific patterns.  All of this adds up to
make peaceful monsters a plague and a nuissance.