Don't treat long worm tail segments as collateral targets for axes.
When swinging an axe in a wide or tremendous arc, don't target any
monster that is dead or whose primary position (m.mx/m.my) doesn't
match the collateral target position.  This effectively means that
long worms will only be targetted on their heads.

Previously, there was a bug wherein targetting their tails sometimes
resulted in a "save file is corrupted" impossible in dmonsfree, when
the number of monsters removed didn't match the number pending because
the same long worm had been killed more than once.

While the idea of being able to hit a long worm in multiple places in
one fell swing is interesting, the really correct way to do it would
be to give each tail segment separate hit points, and in any case
ultimately it wouldn't really be a significant boon to axe wielders
(nobody is going to decide to wield an axe or not based on whether
they can hit multiple long worm segments at once) and therefore isn't
worth the fiddling that would be required to get it to work safely.

Thanks to Yasdorian for pointing out the bug.