Bones creation:  be more paranoid about variable initialization.
Mainly, the whynot variable needs to be initialized in all cases.

Without this change, the testbench has segfault problems quite often.
With this change, it can run through a LOT more tests before it runs
into anything it can't handle.

There may be an underlying bug that the testbench was running into,
that was triggering this.  But the problem that this commit fixes
was making it hard to interpret the testbench's output.