Start healers with a collection of herbs.
Instead of one large stack of eucalyptus, healers now start with a
medium stack of eucalyptus, a small stack of wolfsbane, and a small
stack of garlic.  The wolfsbane and garlic are, individually, nowhere
near as thematic for the healer as the eucalyptus is, but taken
together the collection makes sense.  Thanks to hackedhead for the
suggestion.

Also, this commit contains a const warning fix for the recent armor
size changes.