Tweaks for Dr. Jeckyl's Laboratory
When this level was designed, I didn't realize that the level compile
doesn't support specifying random-appearance objects by their
appearance; I also wasn't realizing how small the limit is on the
number of objects a level can specify.  These restrictions caused
several changes to this level, all of which are disadvantageous to
the player, making the level (already not as good to the player as
some of the other options) significantly worse.

To compensate, I am temporarily adding two scrolls of wishing to
the level, as a sop to the player.

Also, I moved a couple of boxes, because I forgot how terrible the NH4
interface for looting is when there's more than one container on the
floor in the same place.

Also, I had neglected to include a magic chest.  So add that.