Fix the problems with the BingeHack castle levels.
It turns out that the place_lregion problem with The Hall of the Titan
King and The Troll King's Dungeons was caused by the mazewalk
junctions not being on even-numbered coordinates, something that ought
to be possible but isn't due to internal details of how the maze
generator works.  So I fixed that by shifting the mazes up one row
so their coordinates work out correctly.

There were, however, additional issues beyond that that triggered if
ROWNO and COLNO were large.  (The worst of these was that mazewalk
went way beyond the intended area).  These issues have also now been
fixed, in the one case by surrounding the level by water instead of
solid rock (water is more thematically appropriate to the level
anyway), and in the other case by giving the maze area a double thick
back wall.

Finally, neither of these levels had a magic chest, so I added that.

So now there are six possible variants of the last level in the
Dungeons of Doom, as follows:
 1. Castle of Yendor (main prize: wand of wishing)
 2. Aladdin's Palace (main prize: magic lamp)
 3. Ken Arnold Memorial Library (main prize: scroll of wishing)
 4. Dr. Jeckyl's Laboratory (main prize: two scrolls of wishing)
 5. Hall of the Titan King (main prize: wand of wishing)
 6. Troll King's Dungeons (main prize: wand of wishing)

Naturally, in addition to the main prize, they all have fairly
significant amounts of other loot as well.