Implement jitterbugs.
The early game had a niche opening for something really fast-moving
(but weak enough not to be unbalanced despite that).  In filling that
niche, I wanted to introduce some interesting mechanic, so jitterbugs
have a passive ability to displace you (that is, switch positions with
you on the map) after you hit them.

They're not a valid polymorph form, for a couple of reasons, but the
main reason is, I didn't feel like implementing the monster-vs-player
side of the passive displacement attack, when it would basically never
be used (because as a polyform, jitterbugs wouldn't ever be used much
in practice anyhow, due to their other limitations).  Relatedly, I
couldn't think of a reasonable way to test the monster-vs-player side
of this, since it would require polymorphing into a jitterbug, which
is really not a practical polyform (too weak, no ability to manipulate
objects, etc) unless I give the ability to another monster (perhaps
temporarily, for testing, but I don't really like the idea of leaving
the code implemented but sitting dead and unusable).  So for now, I
just made jitterbugs not a valid polymorph form.  I don't think anyone
would really want to polymorph into them anyway.

If the passive-displace mechanic proves to be particularly interesting
and popular, I suppose might potentially be persuaded to flesh it out
a bit more and give it to some other monster that's a bit higher level.

The monster-vs-monster aspect of passive displacement hasn't been
implemented yet either, mostly because I'm lazy, and anyway, FIQ is
probably going to get us to the point where we only have to implement
each of these things once, at which point I'll rebase onto his work
and Bob will gain intrinsic avuncularity for everyone.