Apply the sleep spell duration change when hitting players.
FIQ's adjustment applied to both players and monsters in his variant,
because he has merged the codepaths.  Fourk still has distinct
codepaths for spells hitting players and monsters, so if we want the
behavior to match, we have to make the change twice.

Some day I'll rebase onto FIQHack and not have to do this sort of
thing any more, but for now...