Make ghosts invisible.
Aeons ago, NH4 took away their bogus pseudo-invisibility-via-mlet,
which was already not relevant in graphical tilesets, and which could
be configured away on NAO and any other server with its patches.

That's all well and good, because using bad UI as a substitute for a
game mechanic to provide a challenge is not our thing in the NH4
school of thinking.

But it has occurred to me that ghosts really ought to be invisible,
using the actual game mechanic that exists to support invisible
monsters, which was already used by stalkers, black lights, and
anything that gets a wand of make invisible in starting inventory.