Implement random named weapons, for non-cursed positively enchanted
When a random weapon happens to be generated positively enchanted and
non-cursed, it now has a chance (with increasing probability when its
enchantment is higher) to be automatically assigned a name.  Most of
these names will take the form "Weapon Type of Foo", e.g., "Sword of
Justice", "Spear of Glory", "Mace of Virtue", "Sling of Yendor".

These are not artifacts.  They're just ordinary weapons that happen to
be positively enchanted.  The name is meant to potentially draw the
player's attention to the weapon.  Some players habitually curse-test
and try on a lot of armor, hoping to find positively enchanted gear,
and sometimes they do.  But because weapons do not reveal their
enchantment when equipped, finding randomly enchanted weapons is much
harder unless you can cast the identify spell.  This is an attempt to
address that, while also adding (a small amount of) variety.

(cherry picked from commit 8205a946264c088aeef80e07c5cb2a61d92a3cde)

Conflicts:
	doc/changelog-fourk.txt