Carrying the Amulet of Yendor disrupts teleport control.
Because the mysterious force no longer interferes with going up stairs
in Gehennom, it's been possible in Fourk to go up stairs pretty much
every other turn, if you have adequate sources of teleport, by using
horizontal teleport to bypass most levels (apart from the ones that
are non-teleport, of course).

I didn't want to block teleport entirely, because it's an escape
mechanism, and an important one.  But then this halfway measure
occurred to me: the player can still use horizontal teleport to get
out of a sticky situation, but it's not trivial to just skip an
entire dungeon level.