New furniture terrain: benches.  #sit on one to regain 1 HP (per turn).
The workbenches in the Laboratory are now benches, as originally
intended, rather than thrones.

Benches will also be used in the Garden of Temptation, when that
is implemented, which will probably not be until 4.3.0.5 or later.

I am still deciding whether to have them occasionally show up at
random on ordinary dungeon levels.