Make names of super-adult dragons more consistent.
The age modifier now always comes before the color, and the
progression is as follows:
   baby => young => (implied adult) => elder => ancient

So for example a green dragon egg hatches into a baby green dragon; if
it gains enough experience, it grows up into first a young green
dragon, then a green dragon (which is the main adult form), eventually
an elder green dragon, and ultimately an ancient green dragon.

As before, elder and ancient dragons are not generated randomly but
may be encountered in dragon halls, or if a dragon has somehow gained
enough experience.