Temporary alignment record bonus for carried/equipped items
Provide certain races and roles with an alignment record
bonus for carried/equipped items:
 * Archeologists for carried artifacts and unique items
 * Samurai for a wielded katana, yumi, or tsurugi
 * Gnomes, for carried gems, on a log function of value
 * Giants, for carried gold, on a log function of amount
 * Dwarves, elves, and orcs for equipped racial armor
   and weapons.
 * Gnomes for wielded aklys or crossbow.
 * Humans for a worn fedora.

The bonus only applies as long as the relevant items
are carried or equipped, as applicable.

Most of the lines in this commit are purely mechanical
replacement of u.ualign.record checks with a macro that
adds in the bonus.