Second part of achievements: don't offer locked options.
Options, races, and roles that are unlockable but have not been
unlocked will now not be offered to the player.

Unlockable gameplay modes that are just one choice on the list
of modes (currently, challenge mode) are currently not handled,
so they behave as if always unlocked.  I'll clean that up in a
future commit, now that the rest seems to be working as designed.

Also, now that options can be trivially hidden from new players by
making them unlockable, I've gone ahead and added a number of options
that are purely for the added challenge, such as permanent stunning
and confusion and fumbling and so on.  The only reason I forebore to
implement these previously was because it would clutter the options
screen in a way that could be daunting for new players.  This is now
neatly avoided by keeping them locked until the player demonstrates
significant ability to do well at the game.

There are also unlocks in place for several deferred features
that have not yet been implemented, including two roles (Convict
and Druid) and a gameplay mode (one-hitpoint).  These may not be
implemented until the next version, but I went ahead and put in
the unlock tracking for them, while I was doing it.