Don't use uninitialized pointer when mfind_target finds no target.
A variable, mtmp2, was being assigned the output of mfind_target, and
then the returned struct monst * was being used right away under the
assumption that it was necessarily valid and contained a monster; but
in fact mfind_target is allowed to not find a target.  Chaos ensued.

(cherry picked from commit f0cf7c9d21514e277996b9c6e72917074e1b5166)