Water walking boots, their identification state, and walking onto liquids.
Walking onto water no longer requires the moveonly prefix in the case
where the player is wearing type-identified water-walking boots.

Furthermore, walking onto water with the m prefix type-identifies the
water walking boots, meaning you only have to do it once (per game).

If you know (formally) that the boots are fireproof water walking
boots, you also no longer need the moveonly prefix to walk onto lava,
provided you have fire resistance.  (This could potentially leak
information about the identity of an item that grants fire resistance,
such as a ring thereof, but since the player must formally identify
the boots to know that they are fireproof, and on top of that it can
only ever happen in the presence of lava, this seems like too much of
a corner case to worry about.  Without fire resistance, lava does
enough damage to potentially be a real gotcha, so you still need the
moveonly prefix in that case.)

There are almost certainly some remaining cases wherein something that
happens in the game can informally let the player know that the boots
they are wearing are water walking boots, without formally marking
that fact as known.  But the most common cases are handled, including
the really important one of using the m prefix to step onto water.