Monster throws potion while unaware: change panic to impossible
I'm not sure exactly what's going to blow up, but what could possibly
go wrong?

Realistically, the odds that I will track down the _cause_ of this
bug, deep in the bowels of the AI code that I haven't even started to
understand, in a timely fashion, are nil.  If we can get past this one
instance and it doesn't happen often, I may just table the problem
until I can rebase on FIQ's muse rewrite, which will obviate the issue
entirely.