Fire Brand and Frost Brand differentiation.
This commit makes Fire Brand and Frost Brand more distinctive.
The Brands are now the only artifact short swords in the game.

Additionally, there is a special exception to the rule against
wielding an artifact in the off-hand, second to another weapon:
either of the Brands can be wielded second to the other.

When attacking while wielding two weapons, you now get artifact
damage bonus, when applicable, for the off-hand weapon as well
as for the main-hand one.  Currently, the only way for this to
happen is with the Brands.

I did NOT change how wielded-artifact intrinsics work: you still can
only get one of the two resistances (fire resistance from wielding
Fire Brand in the main hand, or cold resistance from Frost Brand in
the main hand, not both).  This is probably not significant to balance
anyway (because by the time most players see an artifact, they
typically already have both fire and cold resistance intrinsically),
but I like the fact that it does technically matter, at least in
some corner case, which weapon is in which hand.