Allow Sting to be held second to Orcrist.
The player can now wield Orcrist in the main hand and Sting in the off
hand.  Both of these weapons are of such limited usefulness that this
cannot possibly disrupt balance much.  It wouldn't have been worth
doing the exception-to-the-rules infrastructure just for this, but
since I was twinning the Brands anyway, throwing this in too is easy.