When dual-wielding a non-weapon, say you "bash" rather than "hit".
For example, if as a samurai you dual-wield your yumi with your
washizaki and start attacking things in melee combat, you will "stab"
things when the sword hits (this is the same as before) but will now
be said to "bash" things when the yumi hits.

Thanks to hackedhead for noticing that you got "hit" messages
for both, previously.