New room-and-corridor topology, minrand.
Add a new topology option for room-and-corridor levels, wherein the
minimum number of random-room-to-random-room connections are created
so that every room is directly or indirectly connected to every other
room.  With this topology, redundant connections only occur when the
corridors cross each other (which does happen quite a bit).  This
topology is not as hardwired, and thus not as artifically distinctive,
as the ring and hub-and-spoke topologies, but it nonetheless manages
to feel different from the traditional/standard topology.  Corridors
in this topology have a greater tendency to go around the back sides
of rooms, cross one another diagonally, and run at strange angles to
make their connections.