Don't generate deep rock dwellers in grossly inappropriate places.
A player encountered deep ettins on the Plane of Water, which is
clearly not something that should happen.  This change to rndmonst
should hopefully prevent that sort of thing.

Deep rock dwellers (anything with an mlet of S_XORN, i.e., X) should
no longer randomly generate on Planes other than Earth, nor on early
dungeon levels, nor on non-rocky or outdoor-type levels (including a
number of the role quests and certain other special levels).