Key differentiation: main commit, introduces several kinds of keys.
There are now three basic types of keys:
 * keys that lock/unlock chests and boxes (brass key, silver key)
 * keys that lock/unlock doors (bronze key, door key)
 * universal keys that lock/unlock both (iron key, skeleton key, sturdy key)

In addition to that, some keys are subject to damage due to the type
of material they are made from, and some are not.  Brass and bronze
keys, being made of cuprous alloys, are subject to corrosion.  Iron
keys rust.  Skeleton keys, being made of bone, can break outright (as
can non-artifact credit cards, unless you are a Tourist).  On the
other hand, silver keys, (METAL) door keys, and (mithril) sturdy keys
do not rust or corrode.

Rust and corrosion actually matter for keys now:  the more corroded a
key is, the less reliable it is for opening locks quickly.  (It can
still be made to work: it just takes longer.)  The same is also true
of lockpicks, while I was at it.

All of this makes the sturdy key, which can open any lock and will
neither break nor rust nor corrode, the best key the player can have.
But you can get by with a brass key and a bronze key, in a pinch.

Keys that are made of a visually-obvious material such as brass or
silver are trivial to identify, so the game just tells you what they
are.  Iron keys, (METAL) door keys, and (mithril) sturdy keys all look
the same ("key" when unidentified).  Door keys can be distinguished by
the fact that they don't work on boxes, and iron keys by the fact that
they can rust.  Sturdy keys can be distinguished from the others by
price or by weight.  Formal identification also works, of course.

I haven't done graphical tiles support for this yet.