Holiday-themed fruit names and related special-occasion support.
We already had special fruit names for Christmas ("candy cane", "sugar
plum", etc.), as part of the advcal patch.  This extends that by
adding special fruit-name treats for a handful of additional dates.

Also, on the anniversary of the original public NetHack Fourk
announcement, the extra items generated on early dungeon levels are
different items than they would be on any other day.