Allow a wider-range of containers to sometimes protect their contents from water damage.
Allow bags of holding, large boxes, and chests to sometimes (always if
blessed, frequently if uncursed, only occasionally if cursed) protect
their contents from water damage.

Oilskin sacks are still more reliable, but this makes a large box
better than a sack (except for the fact that it weighs more).