Add a utility, tiletool.pl
This simple Perl script can parse the source for the palette-based
pixel tilesets (currently, Slash'EM and DawnLike) and, assuming your
terminal supports 24-bit color, display specific tiles that you ask
for in your terminal.  It's faster than going into the game and
wizwishing up a bunch of items and laying them out on the floor, and
it won't run into problems with certain appearances not being used in
your current game.