DawnLike tileset:  add new appearances for potions, rings, spellbooks
Add the following tiles to the dawnlike-16 tileset:

Potions:
 viscous potion, sticky potion, foaming potion, creamy potion,
 gray potion, deep blue potion, muddy potion, icy potion, amber
 potion, violet potion, aromatic potion

Books:
  dark spellbook, black spellbook, wide spellbook, long spellbook,
  spotted spellbook, striped spellbook, spiral-bound spellbook,
  tattered spellbook, torn spellbook, calligraphed spellbook, scrawled
  spellbook, hand-printed spellbook, woodcut spellbook, embossed
  spellbook, illuminated spellbook, paperback spellbook, stapled
  spellbook, pocket spellbook, pale spellbook, maroon spellbook

Rings:
  plain ring, ridged ring, glass ring, obsidian ring, wedding ring,
  class ring, signet ring, scarab ring, intaglio ring, dime store
  ring, mood ring, promise ring, puzzle ring, bone ring, ceramic ring,
  tin ring

These are intended to be used as extra random item appearances, so
that the set of appearances is not the same every game.  Before that
can happen, of course, the other tilesets will need to support them
as well.

A few of these appearances are already supported by the Slash'EM
tileset, due to being used for items that exist in Slash'EM but not in
vanilla NetHack; but there are also a number of new appearances here.