Add a number of "spare" random appearances for rings, potions, spellbooks.
The ring, potion, and spellbook object classes now have more random
appearances than objects, so (as was already the case with scrolls,
even in vanilla) not every appearance is used every game.

Additionally, certain ring appearances provide information that can be
read: the year on class rings (which is purely for flavor), the
engraving under the stone on intaglio rings or on the face of signet
rings (these are mostly flavor but occasionally show an alchemy
formula), and the color of the stone on mood rings (which works as in
Slash'EM, except that we treat mood ring as an appearance, rather than
the ring's primary function).  Except for mood rings, any given ring
will produce the same information each time it is read (the same way
T-shirts work).