Update rltiles-32 to support the various levels of dragons.
To avoid forking a binary .png file, the new potion, scroll, and
spellbook appearances from the bleeding-edge branch are also included
here, as well as the banshee tile, despite the fact that they won't
actually be used in 4.3.0.4.

But the important thing here is that the RL-Tiles 32 tileset now has
distinct images for five levels (baby, young, (adult), elder, and
ancient) for each major color of dragon.

This update also includes the (currently-unused) banshee tile.

(Portions of this commit are cherry picked from commit
5cab2a8b10b29ee162cd32505cf289043abd29a8)

Conflicts:
	doc/changelog-fourk.txt
	tilesets/dat/tiles/rltiles-32.png
	tilesets/dat/tiles/rltiles.map