Make Izchak willing to buy the Candelabrum of Invocation.
In the 3.6 series, the Dev Team chose to go a different direction,
having him suggest that you keep it and buy candles to go with it.
That's not necessarily BAD, but I couldn't think of any reason why
he wouldn't just buy it, same as any other lighting-related item.

It's a corner case anyway.