Implement additional properties for (most colors of) dragon scales.
Infrastructurally, any object can now provide two properties.
The special-casing on alchemy smocks has been subsumed by this,
and additional properties have been added to (most colors of)
dragon scales and scale mail.  See the changelog entry for a list.

The only _new_ property here is protection from water damage, which
extends to your inventory.  (If anybody asks about the flavor
justification for this:  the water that would otherwise harm your
stuff is turned to tiny ice particles that fall harmlessly away before
melting again.)

On top of that, red and white scales and scale mail now provide
the wearer with the corresponding breath attack.

The actual reason for implementing all of this is, of course, that we
want players to be making more meaningful choices, and in vanilla the
only colors anybody ever seriously considers are silver and gray.  I
am hoping that these extra properties will entice some players to
consider choosing some of the other colors on occasion.