Dragon scales and scaled armor, per dtsund proposal (with variations), main commit.
Dragon scales are now worn in the cloak slot and contribute nothing to
AC; they only provide properties.

Dragon scale mail is no longer a specific type of armor.  Instead, any
body armor can be dragon-scaled.  So for example you can now have
green-scaled orcish ring mail, an orange-scaled elven mithril coat,
red-scaled studded leather armor, or silver-scaled bronze plate mail.

The scales make the armor impervious to erosion damage (if it wasn't
already) and cause it to provide whatever properties the scales
themselves would have provided.  The most important advantage of
combining the scales onto the armor is that it frees up the cloak slot
for another cloak (or another color of scales).

Dragon-scaled armor cannot be wished for directly, but you can wish
for the scales if you want.  Wearing the scales over the armor that
you wish to scale, read a non-cursed scroll of enchant armor (while
not confused), and the armor will acquire absorb the scales, acquiring
an appropriately-colored sheen.

The original proposal suggested being able to remove the scales again
with a tin opener.  This has not been implemented.  Also, dtsund has
said that he meant for it to be possible to apply the scales to the
armor without using a scroll of enchant armor; this has not been
implemented either, at this time.

Scales themselves cannot be meaningfully enchanted.  Attempting to do
so while not wearing body armor will polymorph the player into that
type of dragon.  Attempting to enchant scales over already-scaled
armor will change the armor's scale color; this also may not be quite
what dtsund originally intended, but it'll do for now, IMO.