Runics, Part One: infrastructure for allowing objects to have a rune.
This is the code for generating and maintaining runics on objects, and
for allowing the player to read runed objects (which doesn't
necessarily tell you what they do, if the rune is not identified).

Currently, the runes don't actually do anything.  That'll be a
separate commit, but I want to get this out of my git stash now.