Temple prices scale with the number of times you have purchased protection.
The formula for how much you need to donate to a temple to get a
particular result now scales not only with experience level, but also
with how many times you have donated and received protection already.

To avoid making players consult the changelog or notes every time they
go to a temple, the priest suggests an amount that would be adequate
to receive protection.

Also, the maximum number of points of protection you can get in this
manner is now ten, and the diminishing probability formula is gone.
Hauling tens of thousands of zorkmids up out of Gehennom, trip after
trip, is just tedius, and I don't want players to feel like they need
to do it because it *might* get them a few more points of AC.