Bugfix: when initializing object probabilities, handle extra descriptions correctly.
Adding extra descriptions for rings (commit 2d2f6617316) violated
an implicit assumption in the code that initializes their object
probabilities (which are all 0 in objects.c, signaling that they
should all be equal).  In order to keep the extra descriptions,
we either had to assign the ring probabilities manually (leaving
the implicit assumption as a trap for later) or enhance the code
that initializes the object probabilities so that it doesn't make
that assumption.  I chose the latter course.