By default, don't build the net client.
There are currently no public servers of the type that the net client
would connect to.  There are no immediate plans to have such a server.
The upstream project leader has indicated that the relevant API has
known vulnerabilities and that it would be inadvisable currently to
host such a server, until someone does a security audit and overhaul.

So why are we building the net client every time we build the game?
Let's not.

The net client can still be built by defining the NETCLIENT flag.
It's just no longer the default.