Stronghold fix: don't infinite loop, arriving from below, if narrower than COLNO - 7 columns.
Because of the way in which the arrival-from-below region for the
Castle was hardcoded, the Laboratory, being only 60 columns wide, was
causing an infinite loop any time anyone arrived there from below.

This is a temporary workaround.  What I really want to do is NOT
hardcode the level's arrival region.  But this will at least stop
the infinite loop.