Fix off-by-one error with the non-death game-ending-reason constants.
When I added DISINTEGRATED to the list of possible death reasons, I
failed to notice that the subsequent non-death game-ending reasons
were the next numbers and shouldn't overlap with them.  I left TRICKED
at the same number as GENOCIDED, which is wrong.  This resulted in
people getting the wrong game-ending reason in a logfile in some cases
(though the dumpfile was correct, because apparently the codebase is
not consistent about how it reports these things).