TODO list maintenance and a comment fix for available spell letters.
Thanks to FIQ for pointing out that the comment listed g as being
currently-in-use as a default spell letter, but it was not actually
being used as one.

Move some TODO items to next release, and I went ahead and dealt
with one that had to do with my local build setup here.