Blessed scroll of wishing: don't always allow 127 wand charges.
Thanks to FIQ for noticing this.

To allow for easy testing, it was also necessary to add a scroll of
wishing to the explore mode equipment.  (Debug mode wishes, of course,
are even higher-quality than blessed-scroll wishes and _will_ in fact
always allow 127 wand charges; this is by design.  Having made this
change to explore mode, I left it in, at least for now.  Why not?)

So now you can (with a blessed scroll) wish for 20 charges on a wand,
or (1+rn2(5)) times as many as the wand naturally generated with,
whichever is more.  This makes it theoretically possible (but not
guaranteed) that you can get, for example, a wand of light with 50
charges, from a blessed-scroll wish.  Though I don't know if players
will want to use the scroll that way.