Allow invocation ritual to modify terrain closer to the map edge.
The existing code's limits on how close to the edge the map could be
modified were based on the assumption that the level was a maze level,
which was true in vanilla.  NetHack maze levels can't use the two
tiles closest to the edge of the map, due to limitations in the game's
implementation of the maze walking algorithm (having to do with how
the arithmetic checks are done to see whether a given tile is a
junction or not, and the fact that all junctions must be surrounded by
wall on all sides for code safety reasons).  Pursuant to this, the
invocation code was refusing to modify tiles on or adjacent to where
the edge of the maze would be, so as to avoid going beyond the edges
of the walkable portions of the maze.

In fourk, the vibrating square level is no longer a maze; it's an
aiscav level.  The aiscav algorithm does not share the maze
generator's limits on getting close to the edge of the map, and so we
can modify tiles two rows and/or columns closer to the edge of the map
without causing problems.  The code that decides where to place the
vibrating square on aiscav levels was assuming this, which was leading
to situations where the placement of the vibrating square resulted in
the invocation area being incomplete (and the impossible message
triggering) quite unnecessarily.

So now we calculate whether we can and should change a tile, when
creating the invocation area, based on how close we are to the edges
of the level, rather than how close we are to where the edges of the
maze would be if the level were a maze, which it isn't.