Make crystal balls more usable.
Give crystal balls more charges, and nix the intelligence check.

In the long term, I'd kind of like to make intelligence relevant
again, by using it for the search radius, so that high intelligence
would be needed to search an entire level, but the ball could still be
used by lower-intelligence characters, just with a more limited range.
This is deferred to 4.3.0.5 due to the refactoring that would be
required in order to accomplish it: it's too late in the 4.3.0.4
development cycle to be rewriting the entire logic of several search
functions that, in addition to being used by crystal balls, are also
used for other purposes.