Polymorphing into a python, you now "slither out of" armor.
Previously, you broke out of your armor, but pythons aren't THAT big
around.  (I suppose I could make it still break the armor if you're
bloated beyond a certain point, but so far I haven't bothered.)

Rather than check for python specifically, I checked whether the
target form is slithy but not humanoid and not huge (or larger).  Long
worms for example are gigantic, so polymorphing into one of those will
still break your armor.  (This is correct: in the source material,
those things are bigger around than a human is tall.)

Thanks to raisse for the suggestion on this.