Replace the wiztower-2 zoo with a dragon hall.
An ordinary zoo at that stage of the game is just ridiculous.
The dragon hall is probably still not much of a threat, so
further refinements may be wanted later, but this is at least
a very small step in hopefully a reasonable direction.

I mean, it's either this or manually stock the room with the
nastiest combinations of monsters I can think of, and that
could get a bit too Evil-Patch-ish.