Khortown changes to accommodate the level's new home in Gehennom.
I'm not sure where this level was originally intended to be placed, or
even whether the author had a specific dungeon location in mind.  In
Fourk I have elected to place it in upper Gehennom as an alternative
to the Garden of Temptation.  To accommodate this, I have adjusted the
alignment of the temple altar (it's now aligned to Moloch, as other
altars in Gehennom are) and the statues (they are now of monsters that
make sense in a Garden of Temptation context).  Additionally, I added
a magic chest, because the other Garden levels have one.