Mouse support default to off.
Having mouse support turned on causes various problems, not least, any
time the mouse twitches even slightly, the terminal may long streams
of garbage.  I've personally been turning it off since about two
minutes after I started playing NetHack 4.

I'm leaving the option in, because some players may actually feel that
it's worth it.  But having it be the default is just not reasonable,
especially when people watching someone play on a public server are
going to end up with _their_ terminal's mouse support turned on as a
result.  If the player is actually using the mouse and so turned the
feature on deliberately, I'm willing to live with that.  But having
this happen to watchers just because the player neglected to bother to
turn mouse support off even though they aren't using it is very
annoying.