When lregion cannot be placed, try to place it anyway.
lregions are important; they tend to be things like stairs, arrival
regions, etc., and so failure to generate them introduces unintended
discontinuities in the topology of the entire dungeon.  This, IMO, is
more important than perfectly preserving the design of an individual
level, even if it's a special level.

So if there is initially no place to put an lregion under the
preferred set of restrictions, throw away the least important
of those restrictions and try again.

When this happens in debug mode, a debug pline() message is issued
for each restriction that is discarded.